requirements.txt
python-amazon-paapi==4.2.1
amightygirl.paapi5-python-sdk==1.0.0
python-telegram-bot==13.4.1
requests==2.25.1
